+++ src/main/java/com/redcom/redcom/Exceptions/UserException.java
@@ -1,0 +1,22 @@
+package com.redcom.redcom.Exceptions;
+
+public class UserException extends RuntimeException {
+
+    private String message;
+
+    public UserException(String message)
+    {
+
+        this.message = message;
+    }
+
+    public String getMessage()
+    {
+        return message;
+    }
+
+    public String setMessage(String message)
+    {
+        return this.message = message;
+    }
+}
