+++ src/main/java/com/redcom/redcom/Exceptions/UserExceptionHandler.java
@@ -1,0 +1,18 @@
+package com.redcom.redcom.Exceptions;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class UserExceptionHandler {
+
+    @ExceptionHandler(UserException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public UserException handlUserException(UserException ue)
+    {
+        return ue;
+    }
+
+}
