+++ src/main/java/com/redcom/redcom/controllers/HelloController.java
@@ -1,10 +1,12 @@
-package com.redcom.redcom;
+package com.redcom.redcom.controllers;
 
 import java.util.concurrent.atomic.AtomicLong;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.redcom.redcom.dto.Hello;
 
 @RestController
 public class HelloController {
