+++ src/main/resources/application.properties
@@ -1,1 +1,3 @@
 server.port = 9085
+
+spring.jackson.serialization.indent_output = true
